Allow for hiding of children on a primary navigation item
Some primary navigation menu items will not have subnavs, therefore,
we add the option for site Admins to ensure that for such menu items,
the children are not displayed in the navigation menu.

This is done by extending the existing LinkBlock.
In order to make things cleaner and ensure uniformity across the
codebase, all the blocks in the navigation app are moved to their own
blocks.py module.