Fix an issue with Google Chrome
Without this, attempting to play a video in Google Chrome
will result in some weird behaviour, where the video
doesn't immediately play, forcing you to click the play button
again, which is not a pleasant UX.

While this seems like we are setting videos to autoplay
when the page loads, this is not the case. With the
lite-youtube-embed component, the video NEVER even loads
until one clicks the play button.